python/requirements.txt
Flask==2.0.1
requests==2.26.0
tabula-py==2.3.0
Pillow==8.2.0
numpy==1.22.0
opencv-python==4.9.0.80
pytesseract==0.3.10
img2table==1.2.8
Werkzeug==3.0.1
Flask-Cors==4.0.0
gunicorn==21.2.0
pika==1.3.2
psycopg2==2.9.9
asyncio==3.4.3
websockets==12.0
PyMuPDF==1.23.14
# fitz
# tqdm
firebase-admin==6.3.0